+++ custom_components/pstryk/config_flow.py
@@ -1,5 +1,5 @@
 # Marcin Koźliński
-# Ostatnia modyfikacja: 2026-03-28
+# Ostatnia modyfikacja: 2026-03-29
 
 """Config flow for Pstryk Energy integration."""
 
@@ -94,7 +94,7 @@ class PstrykConfigFlow(ConfigFlow, domain=DOMAIN):
                     vol.Optional(CONF_TIMEZONE, default=DEFAULT_TIMEZONE): str,
                     vol.Optional(
                         CONF_SCAN_INTERVAL_MINUTES, default=DEFAULT_SCAN_INTERVAL
-                    ): vol.All(vol.Coerce(int), vol.Range(min=15, max=120)),
+                    ): vol.All(vol.Coerce(int), vol.Range(min=30, max=120)),
                 }
             ),
             errors=errors,
@@ -138,7 +138,7 @@ class PstrykOptionsFlow(OptionsFlow):
                         default=self.config_entry.options.get(
                             CONF_SCAN_INTERVAL_MINUTES, DEFAULT_SCAN_INTERVAL
                         ),
-                    ): vol.All(vol.Coerce(int), vol.Range(min=15, max=120)),
+                    ): vol.All(vol.Coerce(int), vol.Range(min=30, max=120)),
                 }
             ),
         )